changed aniindex to int

+++ Warlock The Soulbinder/Player.cs
@@ -15,7 +15,7 @@ namespace Warlock_The_Soulbinder
         private bool graceStart = true;
         private int graceSwitch = 0;
         private Vector2 lastDirection = new Vector2(0, 1);
-        private double aniIndex;
+        private int aniIndex;
         private double elapsedTime;
         private const int AnimationFPS = 30;
         private bool walking;
